% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.tinyint.R
\name{as.tinyint}
\alias{as.tinyint}
\alias{as.tinyint.numeric}
\alias{as.tinyint.double}
\alias{as.tinyint.integer}
\title{Convert to \code{tinyint}}
\usage{
as.tinyint(x, ...)

\method{as.tinyint}{numeric}(x, ...)

\method{as.tinyint}{double}(x, ...)

\method{as.tinyint}{integer}(x, ...)
}
\description{
Convert to \code{tinyint}
}
